react-ui, react-ui-components, app-toolkit: one progress readout, adapting to what the monitor reports
`ProgressMeter` collapses into `Progress.Root`. It is one component with one
geometry — three explicit grid rows, always drawn — and it picks its parts from
the state rather than changing shape:

- a bar fraction where the phase counts, a sweep where it cannot;
- the step chain where the run declared a plan;
- a crawl of the phases it has named, and the cancel control.

The old meter dropped the bar row entirely when a phase stopped being countable,
which re-flowed the rows below it into a 4px track. Measured in Chromium across a
scripted run — determinate, uncountable, and back — the readout now holds 56px
and rows 24/4/24 throughout, including on failure.

The bar is the reconciled one: `ProgressMeter`'s colours, width transition and
`aria-valuenow` (a separator track under a primary fill actually contrasts, where
the old base-surface track was invisible against its host), keeping `Status`'s
indeterminate sweep. The dead `main-bottom` variant goes with it — nothing has
passed it since the layout it was named for.

The chain had to keep an intrinsic width: given none it measured zero and drew
nothing. It now caps at the plan's natural size and shrinks from there.

app-toolkit's adapter stays exactly where it was, so its four consumers are
untouched; it just renders `Progress.Root` and no longer reaches through
react-ui-components.
