--- packages/ui/react-ui-components/src/components/Progress/ProgressMeter.stories.tsx
@@ -1,189 +1,0 @@
-//
-// Copyright 2026 DXOS.org
-//
-
-import { type Meta, type StoryObj } from '@storybook/react-vite';
-import React, { useCallback, useEffect, useRef, useState } from 'react';
-
-import { IconButton, Panel, type ProgressState, Toolbar } from '@dxos/react-ui';
-import { withLayout, withTheme } from '@dxos/react-ui/testing';
-
-import { ProgressMeter, type ProgressMeterProps } from './ProgressMeter';
-
-const TICK_MS = 400;
-
-/**
- * One step of a scripted run. A phase either counts (`total` set) or does not — a run that mixes the
- * two is the case the meter has to serve: magazine curation counts its feeds, then makes a single
- * opaque agent call, then counts what it writes.
- */
-type Phase = {
-  note: string;
-  /** Omitted for a phase whose length is unknowable. */
-  total?: number;
-  /** How long an uncounted phase runs; a counted one ends when it reaches `total`. */
-  durationMs?: number;
-};
-
-/** A run whose length is known throughout: a labelled bar with a count and an ETA. */
-const DETERMINATE: Phase[] = [{ note: 'Fetching messages', total: 40 }];
-
-/** A run that counts what it can and cannot count the rest — the case the meter exists for. */
-const INDETERMINATE: Phase[] = [
-  { note: 'Syncing feeds', total: 5 },
-  { note: 'Selecting articles', durationMs: 8_000 },
-  { note: 'Adding to magazine', total: 3 },
-];
-
-const IDLE: ProgressState = {
-  label: 'Curating Reading List',
-  current: 0,
-  status: 'pending',
-  elapsedMs: 0,
-  note: 'Idle',
-  cancellable: true,
-};
-
-/**
- * Drives a task through a scripted run, so the meter is watched from zero rather than joined halfway
- * — the only way to see the elapsed clock, the phase transitions and the cancel control do their job.
- *
- * The toolbar starts either shape and forces a failure; the meter's own ✕ cancels.
- */
-const DefaultStory = (args: ProgressMeterProps) => {
-  const [state, setState] = useState<ProgressState>(IDLE);
-  const [script, setScript] = useState<{ phases: Phase[]; run: number }>({ phases: [], run: 0 });
-  const stopped = useRef(false);
-
-  useEffect(() => {
-    if (script.run === 0) {
-      return;
-    }
-
-    stopped.current = false;
-    const startedAt = new Date().toISOString();
-    const start = Date.now();
-    let index = 0;
-    let count = 0;
-
-    const patch = (next: Partial<ProgressState>) =>
-      setState((prev) => ({
-        ...prev,
-        ...next,
-        startedAt,
-        elapsedMs: Date.now() - start,
-      }));
-
-    const interval = setInterval(() => {
-      if (stopped.current) {
-        return;
-      }
-
-      const phase = script.phases[index];
-      if (!phase) {
-        clearInterval(interval);
-        patch({ status: 'done', note: 'Done', phase: script.phases.length - 1 });
-        return;
-      }
-
-      // A counted phase advances until it reaches its total; an uncounted one only runs the clock,
-      // which is the point — there is nothing to count, so the meter shows elapsed instead.
-      if (phase.total !== undefined) {
-        count += 1;
-        patch({ note: phase.note, phase: index, total: phase.total, current: count });
-        if (count >= phase.total) {
-          index += 1;
-          count = 0;
-        }
-      } else {
-        count += TICK_MS;
-        patch({ note: phase.note, phase: index, total: undefined, current: 0 });
-        if (count >= (phase.durationMs ?? 3_000)) {
-          index += 1;
-          count = 0;
-        }
-      }
-    }, TICK_MS);
-
-    return () => clearInterval(interval);
-  }, [script]);
-
-  const handleStart = useCallback((phases: Phase[]) => {
-    stopped.current = true;
-    // Stamped here, not in a fixture: the clock has to read from zero on start, and a baked-in
-    // `startedAt` shows a run that was already going before anyone pressed anything.
-    // `phases` is the plan's length; the meter draws one circle per phase and marks the one in flight.
-    setState({ ...IDLE, status: 'running', startedAt: new Date().toISOString(), phases: phases.length, phase: 0 });
-    setScript(({ run }) => ({ phases, run: run + 1 }));
-  }, []);
-
-  /** Cancelling is not a failure: the run simply stops, and the meter returns to where it began. */
-  const handleCancel = useCallback(() => {
-    stopped.current = true;
-    setState({ ...IDLE });
-  }, []);
-
-  const handleFail = useCallback(() => {
-    stopped.current = true;
-    setState((prev) => ({
-      ...prev,
-      status: 'error',
-      error: 'Network unreachable',
-    }));
-  }, []);
-
-  return (
-    <Panel.Root>
-      <Panel.Toolbar asChild>
-        <Toolbar.Root>
-          <IconButton icon='ph--play--regular' label='Determinate' onClick={() => handleStart(DETERMINATE)} />
-          <IconButton
-            icon='ph--dots-three-outline--regular'
-            label='Indeterminate'
-            onClick={() => handleStart(INDETERMINATE)}
-          />
-          <IconButton icon='ph--warning--regular' label='Fail' onClick={handleFail} />
-          <IconButton icon='ph--x--regular' label='Cancel' onClick={handleCancel} />
-        </Toolbar.Root>
-      </Panel.Toolbar>
-      <Panel.Content />
-      <Panel.Statusbar asChild>
-        {/* The meter's own control cancels a run in flight, and clears one that failed. */}
-        <ProgressMeter {...args} state={state} onCancel={handleCancel} />
-      </Panel.Statusbar>
-    </Panel.Root>
-  );
-};
-
-const meta = {
-  title: 'ui/react-ui-components/ProgressMeter',
-  component: ProgressMeter,
-  render: DefaultStory,
-  decorators: [withTheme(), withLayout({ layout: 'column' })],
-  parameters: {
-    layout: 'fullscreen',
-  },
-} satisfies Meta<typeof ProgressMeter>;
-
-export default meta;
-
-type Story = StoryObj<typeof meta>;
-
-/**
- * **Determinate** counts to a known total: a bar, `n / total`, and an ETA.
- *
- * **Indeterminate** mixes both — it counts 5 feeds, then cannot count the agent call, then counts
- * again. The bar gives way to an elapsed readout where there is nothing to count, and the phase is
- * named underneath throughout: `label` holds the run's identity and `note` the phase, so the task
- * never reads as a different task mid-run.
- *
- * A run that declares `phases` draws one circle per phase, with the one in flight ringed — the one
- * thing an uncountable phase can still report. **Fail** ends the run with an error. **Cancel** — in the toolbar, or the meter's own ✕ — stops a
- * run and returns the meter to idle: cancelling is not a failure, so it leaves no error behind. The
- * ✕ stays available on a failed run, where it clears the error rather than cancelling anything.
- */
-export const Default: Story = {
-  args: {
-    state: IDLE,
-  },
-};
