--- packages/ui/react-ui-components/src/components/Progress/index.ts
@@ -1,5 +1,0 @@
-//
-// Copyright 2026 DXOS.org
-//
-
-export * from './ProgressMeter';
