+++ packages/ui/react-ui-components/src/components/index.ts
@@ -6,7 +6,6 @@ export * from './AnimatedBorder';
 export * from './HtmlViewer';
 export * from './Matrix';
 export * from './NumericTabs';
-export * from './Progress';
 export * from './QueryEditor';
 export * from './QueryForm';
 export * from './Shimmer';
